chore: Updated tests
