+++ tests/ember-intl/tests/integration/helpers/format-message/input-is-falsy-test.ts
@@ -1,4 +1,4 @@
-import { render, resetOnerror, setupOnerror } from '@ember/test-helpers';
+import { render } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 import { setupIntl } from 'ember-intl/test-support';
 import { module, test } from 'qunit';
@@ -10,7 +10,7 @@ module(
     setupRenderingTest(hooks);
     setupIntl(hooks, 'en-us');
 
-    test('input is null and allowEmpty is true', async function (assert) {
+    test('input is null', async function (assert) {
       await render(hbs`
         <div data-test-output>
           {{format-message null}}
@@ -20,25 +20,14 @@ module(
       assert.dom('[data-test-output]').hasText('');
     });
 
-    test('input is undefined and allowEmpty is true', async function (assert) {
-      setupOnerror((error: Error) => {
-        assert.strictEqual(
-          error.message,
-          '{{format-message}} helper requires a value.',
-        );
-
-        assert.step('ember-intl throws an error');
-      });
-
+    test('input is undefined', async function (assert) {
       await render(hbs`
         <div data-test-output>
           {{format-message undefined}}
         </div>
       `);
 
-      assert.verifySteps(['ember-intl throws an error']);
-
-      resetOnerror();
+      assert.dom('[data-test-output]').hasText('');
     });
   },
 );
