+++ tests/ember-intl/tests/integration/helpers/t/input-is-not-well-defined-test.ts
@@ -1,6 +1,6 @@
-import { render } from '@ember/test-helpers';
+import { render, resetOnerror, setupOnerror } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
-import { setupIntl, t } from 'ember-intl/test-support';
+import { addTranslations, setupIntl, t } from 'ember-intl/test-support';
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'test-app-for-ember-intl/tests/helpers';
 
@@ -10,7 +10,60 @@ module(
     setupRenderingTest(hooks);
     setupIntl(hooks, 'en-us');
 
-    test('input is not a valid translation key', async function (assert) {
+    test('input is null', async function (assert) {
+      await render(hbs`
+        <div data-test-output>
+          {{! @glint-expect-error }}
+          {{t null}}
+        </div>
+      `);
+
+      assert.dom('[data-test-output]').hasText('t:null:()');
+    });
+
+    test('input is undefined', async function (assert) {
+      await render(hbs`
+        <div data-test-output>
+          {{! @glint-expect-error }}
+          {{t undefined}}
+        </div>
+      `);
+
+      assert.dom('[data-test-output]').hasText('t:undefined:()');
+    });
+
+    test('input is an empty string', async function (assert) {
+      addTranslations('en-us', {
+        '': 'Hello, world!',
+      });
+
+      setupOnerror(() => {
+        assert.step('@formatjs/intl throws an error');
+      });
+
+      await render(hbs`
+        <div data-test-output>
+          {{t ""}}
+        </div>
+      `);
+
+      assert.verifySteps(['@formatjs/intl throws an error']);
+
+      resetOnerror();
+    });
+
+    test('input is false', async function (assert) {
+      await render(hbs`
+        <div data-test-output>
+          {{! @glint-expect-error }}
+          {{t false}}
+        </div>
+      `);
+
+      assert.dom('[data-test-output]').hasText('t:false:()');
+    });
+
+    test('translation does not exist', async function (assert) {
       await render(hbs`
         <div data-test-output="1">
           {{t "key.does.not.exist"}}
@@ -18,18 +71,6 @@ module(
       `);
 
       assert.dom('[data-test-output="1"]').hasText(t('key.does.not.exist'));
-    });
-
-    test('translation is not defined for the locale', async function (assert) {
-      await render(hbs`
-        <div data-test-output>
-          {{t "smoke-tests.parent.child" locale="fr-fr"}}
-        </div>
-      `);
-
-      assert
-        .dom('[data-test-output]')
-        .hasText(t('smoke-tests.parent.child', { locale: 'fr-fr' }));
     });
   },
 );
